Fix type checking and import errors from test fix plan 007
- Fix DBException conditional imports with proper fallbacks
- Fix migration.py to handle command.revision returning list of Scripts
- Fix asyncpg timeout parameter type conversion (float to int)
- Fix async/sync callback mismatches in test files
- Fix AsyncEngine/sessionmaker compatibility issues
- Fix test files using non-existent methods and classes
- Add mypy wrapper script to handle transformers internal error
- Update check_code_quality.sh to use mypy wrapper
- Fix ProcessingConfigRequest parameter usage in tests
- Fix type annotations for conditional imports

These changes improve Cursor IDE compatibility and reduce type checking errors significantly.

+++ app/services/metrics_collection.py
@@ -1077,11 +1077,11 @@ class MetricsCollectionService:
             system_metrics = list(self._system_metrics_buffer)
             aggregated = {"count": len(system_metrics)}
             if system_metrics:
-                aggregated["avg_cpu_usage"] = mean(
-                    [m.cpu_usage for m in system_metrics]
+                aggregated["avg_cpu_usage"] = float(
+                    mean([m.cpu_usage for m in system_metrics])
                 )
-                aggregated["avg_memory_usage"] = mean(
-                    [m.memory_usage for m in system_metrics]
+                aggregated["avg_memory_usage"] = float(
+                    mean([m.memory_usage for m in system_metrics])
                 )
         else:
             return
